chore: update config imports
