+++ packages/vitest/src/node/workspace/resolveWorkspace.ts
@@ -3,9 +3,8 @@ import { isMainThread } from 'node:worker_threads'
 import { dirname, relative, resolve } from 'pathe'
 import { mergeConfig } from 'vite'
 import fg from 'fast-glob'
-import type { UserWorkspaceConfig, WorkspaceProjectConfiguration } from '../../public/config'
 import type { Vitest } from '../core'
-import type { UserConfig } from '../types/config'
+import type { UserConfig, UserWorkspaceConfig, WorkspaceProjectConfiguration } from '../types/config'
 import type { WorkspaceProject } from '../workspace'
 import { initializeProject } from '../workspace'
 import { configFiles as defaultConfigFiles } from '../../constants'
